review tests, remove unnecessary/dead ones
